refactor: common components 폴더 구조 변경
