+++ src/components/Layout/Layout.tsx
@@ -4,7 +4,7 @@ import Sns from "@/assets/svg/sns.svg?react";
 import Album from "@/assets/svg/album.svg?react";
 import Who from "@/assets/svg/who.svg?react";
 import ROUTE_PATH from "@/constants/ROUTE_PATH";
-import Navbar from "@/components/common/Navbar";
+import Navbar from "@/components/common/navbar/Navbar";
 
 const GNB = [
   { icon: <Home />, label: "홈", path: ROUTE_PATH.HOME },
