+++ src/components/common/drawer/settings/delete/DeleteImage.tsx
@@ -1,4 +1,4 @@
-import Button from "../../../Button";
+import Button from "../../../button/Button";
 
 interface DeleteImageProps {
   onClickCancle: () => void;
