+++ src/components/common/tag/index.ts
@@ -1,0 +1,2 @@
+export { default as EmotionTag } from "./EmotionTag";
+export { default as KeywordTag } from "./KeywordTag";
