+++ src/components/common/thumbnail/Thumbnail.stories.tsx
