+++ src/components/common/thumbnail/Thumbnail.tsx
