+++ src/components/common/thumbnail/ThumbnailGrid.stories.tsx
