+++ src/components/common/thumbnail/index.ts
@@ -1,0 +1,2 @@
+export { default as Thumbnail } from "./Thumbnail";
+export { default as ThumbnailGrid } from "./ThumbnailGrid";
