+++ src/components/pages/write/ReviewContent.tsx
@@ -1,4 +1,4 @@
-import EmotionTag from "../../common/EmotionTag";
+import EmotionTag from "../../common/tag/EmotionTag";
 import ImageCarousel from "../diary/diary/ImageCarousel";
 import HeartIcon from "@/assets/svg/heart.svg?react";
 import { DiaryImage } from "../../../types/types";
