+++ src/pages/diary/DiaryFlow.tsx
@@ -5,7 +5,7 @@ import { useEffect, useState } from "react";
 import { useErrorBoundary } from "react-error-boundary";
 import { useBlocker, useNavigate, useLocation } from "react-router-dom";
 import Appbar from "@/components/common/appbar/Appbar";
-import Button from "@/components/common/Button";
+import Button from "@/components/common/button/Button";
 import {
   Dialog,
   DialogClose,
